Small TWI refactors
Allow >255B transactions (not that I expect any in practice), and move
common input parsing logic for TwiCommand into a separate function.

+++ include/twi.h
@@ -1,5 +1,6 @@
 #pragma once
 
+#include <stddef.h>
 #include <stdint.h>
 
 struct Twi final {
@@ -7,8 +8,8 @@ struct Twi final {
    public:
     explicit constexpr Device(uint8_t sla) : sla_{sla} {}
 
-    void read(uint8_t addr, void* buffer, uint8_t length) const;
-    void write(uint8_t addr, const void* buffer, uint8_t length) const;
+    void read(uint8_t addr, void* buffer, size_t length) const;
+    void write(uint8_t addr, const void* buffer, size_t length) const;
 
     // Single-byte utilities
     inline void read(uint8_t addr, uint8_t& byte) const {
